converted output to integer

+++ 295B-WebModule_v_2/DataAnalyzer/src/main/java/edu/sjsu/dataanalyzer/utils/pyAlgorithms.py
@@ -27,18 +27,18 @@ if(sys.argv[1]=='BOOSTED_DECISION_TREE'):
 	regr_2 = AdaBoostRegressor(DecisionTreeRegressor(max_depth=4), n_estimators=300, random_state=rng)
 	trained2=regr_2.fit(train_data[inputColumns].values, train_data[outputColumns].values)
 	y_2 = trained2.predict(test_data[inputColumns])
-	print(y_2)
+	print(y_2.astype(np.int64))
 elif(sys.argv[1]=='DECISION_TREE'):
 	# DECISION TREE
 	regr_1 = DecisionTreeRegressor(max_depth=4)
 	trained1=regr_1.fit(train_data[inputColumns].values, train_data[outputColumns].values)
 	y_1 = trained1.predict(test_data[inputColumns])
-	print(y_1)
+	print(y_1.astype(np.int64))
 elif(sys.argv[1]=='GRADIENT_BOOSTING'):
 	# GRADIENT BOOSTING
 	regr_3 = ensemble.GradientBoostingRegressor(n_estimators=80, learning_rate = .05, max_depth = 10,min_samples_leaf = 20)
 	trained3= regr_3.fit(train_data[inputColumns].values, train_data[outputColumns].values)
 	y_3 = trained3.predict(test_data[inputColumns])
 	#y_4 = regr_3.score(train_data[inputColumns].values, train_data[outputColumns].values,sample_weight=None)
-	print(y_3)
+	print(y_3.astype(np.int64))
 	#print("Yield score using gradient boosting: ",y_4)
